Change arguments of event listeners
It was incorrectly named `currentTarget` while the intended
behavior was to pass on the e.target, or in case of
`on('click a')` usage the DOM node matching the selector.

While this change may be confusing, it should not be breaking
since the code now better reflects what is advertised.